feat: move cli logic to lib.rs - adds tests for args - fixes immediate close problem - fixes hot loop if relayer source offline
