+++ crates/node-cli/Cargo.toml
@@ -24,6 +24,9 @@ tokio = { workspace = true }
 tracing = { workspace = true, optional = true }
 tracing-subscriber = { workspace = true, optional = true }
 
+[dev-dependencies]
+uuid.workspace = true
+
 [features]
 default = ["tracing"]
 tracing = ["dep:tracing", "dep:tracing-subscriber", "essential-node-api/tracing"]
